bump acts-as-taggable-on
