+++ Gemfile
@@ -89,7 +89,7 @@ gem 'omniauth-wordpress','0.2.1'
 
 # Tags
 
-gem 'acts-as-taggable-on', '2.4.1'
+gem 'acts-as-taggable-on', '3.2.6'
 
 # URIs and HTTP
