+++ config/initializers/acts_as_taggable_on.rb
@@ -1,2 +1,2 @@
-require 'models/acts_as_taggable_on/tag'
+require 'models/acts_as_taggable_on-tag'
 ActsAsTaggableOn.force_lowercase = true
